requirements.txt
cryptography>=41.0.0
face_recognition>=1.3.0
opencv-python>=4.8.0
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.21.0
